refactor(search): 使用 InternalExecutorFactory 优化搜索线程池管理
- 通过 InternalExecutorFactory 创建并管理搜索执行线程池和索引同步线程池
- 添加线程池包装器，将 ExecutorService 包装为 ThreadPoolTaskExecutor 以支持优雅关闭
- 保留回退实现以兼容无 common-thread classpath 的场景
- 线程池相关配置改为动态计算线程数量，提升灵活性和资源利用率
- 删除原始线程池创建方法，统一线程池生命周期管理
- 修改部分代码将 double 类型转换为 BigDecimal，提升数值精度处理能力

+++ ydsz-literule/ydsz-literule-web/src/main/java/com/njydsz/literule/web/controller/CEPController.java
@@ -1,5 +1,6 @@
 package com.njydsz.literule.web.controller;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -322,7 +323,7 @@ public class CEPController {
     vo.setRuleCode(p.getRuleCode());
     vo.setName(p.getName());
     vo.setWindow(p.getWindow());
-    vo.setThreshold(p.getThreshold());
+    vo.setThreshold(BigDecimal.valueOf(p.getThreshold()));
     vo.setEventType(p.getEventType());
     vo.setFilter(p.getFilter());
     vo.setDescription(p.getDescription());
@@ -339,7 +340,7 @@ public class CEPController {
             ? null
             : h.getMatchedEvents().stream().map(e -> (Object) e).toList());
     vo.setHitAt(h.getHitAt());
-    vo.setMetric(h.getMetric());
+    vo.setMetric(BigDecimal.valueOf(h.getMetric()));
     vo.setContext(h.getContext());
     return vo;
   }
